Update README.md
The text is the same that we had in the common repository which originally had the two notebooks, now it's adapted to describe only the Python notebook and refer to the R notebook that lives in another repository. We made a split because of the requirements to submit to the Journal of Digital History (only choosing one programming language), we then decided to move out the data preparation part in Python to this repository, and submit only the R analysis notebook for publication.