Create script_dataPreparation_Python_unlinkedArticle.ipynb
This is the first version of the python (data preparation) notebook that was submitted in May 2022. This version of the notebook works with the data versions labelled as 2022-04-14, which had specific column names for indicating the uncertainty. In further versions these column names changed, thus, this notebook is kept as it was to work together with the data it was submitted with in the first place (the graphics and outputs were generated with this version of the notebook). This notebook didn't have proper paths to the data.