Update script_dataPreparation_Python_unlinkedArticle.ipynb
Uncommented line that excludes some errors in the data using the indexes for the Certain slice. I had done this for the slice ALL, but forgot to do it for the certain slice, which was giving errors to Ingeborg: "The problem the scripts gives is that there are duplicated vertices (nodelist) names in the files. I remember that we had this problem before, but I do not remember how we solved this. Do you have any idea why there are duplicates in the csv files (certain)?" (email Feb.17,2023)